+++ common/src/main/cpp/sensors/Encoder.cpp
@@ -1,5 +1,5 @@
 #include "sensors/Encoder.h"
 
-double curtinfrc::sensors::EncoderTranslator::GetEncoderTicks() {
+int curtinfrc::sensors::EncoderTranslator::GetEncoderTicks() {
     return _counterBase.Get();
 }
